requirements.txt
asgiref==3.4.1
Django==3.2.5
django-braces==1.14.0
django-cors-headers==3.8.0
djangorestframework==3.12.4
Pillow==8.3.1
psycopg2==2.9.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
